caja-main.c: Initial port to GtkApplication
In this case, holding the application open the same way it is done in Nemo and Nautilus

Based on code from mate-desktop/engrampa@1c4f7a6 in Engrampa and extended with help from reference to Nautilus and Nemo. Nowhere near complete (libunique seems to be all over Caja!), but nothing crashes or renders abnormally.