caja-application.c: Initial port to GtkApplication
Based on code from mate-desktop/engrampa@1c4f7a6 in Engrampa and extended with help from reference to Nautilus and Nemo. Nowhere near complete (libunique seems to be all over Caja!), but nothing crashes or renders abnormally. In this case, keep code not needed with GtkApplication out of GTK3 builds