caja-application.h: get rid of libunique
Remove libunique calls entirely. Note that although this could have been put behind GTK3 selectors, doing that in caja-application.h caused build failures and in both cases made the code MUCH harder to read and understand.

One issue came from this and probably needs new code from scratch: It appears that one of the _unique_message_data functions is responsible for displaying the home, trash, computer, and (if selected) network icons on the desktop as they do not appear.