Fix command line help broken by prev commits
Stop caja --help from crashing caja, make it actually display the help. Had to make all of it one screen and treat it as a series of strings. The normal help invocation from the command line (before this) works fine if Caja is not running, but if it is it takes over the thread and causes caja to exit. Disable that help option, redirect the command line invocation to display this help screen by other means. 

Note that the session management help is excluded until we know what the session manager is going to do when this is all finished, then it can be added back with the appropriate edits. A kludge but the only way i could save the command line help given GNOME's fix came nearly six months after the commit sequence used to port to GtkApplication.

https://github.com/GNOME/nautilus/commit/5c949fffcb0de62fd5375f546fa94dd50ec7f86f
application: rewrite GApplication command line handling

Unfortunately Nautius 3.0 got all the way out the door with both this issue and another bug (that now also affects caja), both listed below, GNOME's fix thus ended up being to a radically different version of nautilus-application.c and I was unable to apply it even by hand to our version based not on May 2011 but December 2010 Nautilus code.

https://bugzilla.gnome.org/show_bug.cgi?id=637157 (in Nautilus 3.0). 
Invoking "nautilus --help" in terminal crashes existing Nautilus instance

https://bugzilla.gnome.org/show_bug.cgi?id=649336
Invoking  "nautilus ." from any directory opens /home directory 

GNOME's basic description of the first bug and their fix is this:
The  handling of the command line was done in the remote process.
Thus an unknown command like --help would kill the main thread as it was processed remotely.
Thus use the local_command_line handler to process locally the command line and  ignore unknown commands and help request remotely.
In the process fix a bug with the application quit, that is it was using release twice but the refcount could increment more than twice thus an invocation of quit would fail. Use application quit_mainloop instead for the non exit with last window use case.