Fix GTK 2 build
As expected found an error blocking a GTK2 build-but only in this file. Fix it and GTK2 build now completes. Can't test it in my GTK3 MATE environment (get GTK2 and GTK3 symbol mix) but was able to build over a temporary installation of mate-desktop built with GTK2 in /usr/local.