GTK3: fix desktop after sudo spatial window bug
For some reason the or operator || sometimes led to the gsetting preference
(which does not work right in sudo invocation) not being re-read when the
user next opened a window from the desktop(which is a spatial window always).
The result was spatial windows opening from the desktop even when
always use browser was set if and only if sudo caja had been run at least once.

For some reason splitting the loop fixed this, possibly by slowing down read.